Add jvm stats collection
This uses the admin/metrics API endpoint to collect JVM stats.
This does not appear to be available for solr4, but the solr6 docs
do include this endpoint.I'm not sure about solr5. This moves the
http client setup and solr base URL building to main.go so that
both the existing exporter and the jvm exporter can share the config.