Add testing for mbeans processing (#14)
Solr 7 changes the key names in the JSON for mbeans metrics. This adds
testing, in conjunction with the exported JSON from different solr
versions, for mbeans processing that will fail if any version of solr
output doesn't work for the mbeans processing. This can serve as a
framework for writing tests against solr output for different versions
for other metric categories as well.

These test currentl fail for solr 7 because the keys changed to be more
dynamic. I haven't found a good way resolve that yet since dynamic JSON
in golang can be challenging.